docs(quiz): use template conditionally for quiz pages
