+++ docs/_templates/quiz.html
@@ -1,19 +1,8 @@
-<!doctype html>
-<html>
-<head>
-  <meta charset="utf-8"/>
-  <title>Marked in the browser</title>
-</head>
-<body>
+{% extends "base.html" %}
+
+{% block body -%}
+{% include "partials/icons.html" %}
   <div id="quiz-container">
-    <div v-html="htmlContent"></div>
+    <div v-html="quizContent"></div>
   </div>
-  <script src="https://cdn.jsdelivr.net/npm/marked/marked.min.js"></script>
-  <script src="https://cdn.jsdelivr.net/npm/vue@2/dist/vue.js"></script>
-  <script src="./quiz.js"></script>
-  <script>
-    document.getElementById('content').innerHTML =
-      marked('# Marked in the browser\n\nRendered by **marked**.');
-  </script>
-</body>
-</html>
+{%- endblock %}
